feat: initialize core OS components and subsystems with production configurations

+++ kernel/include/sync.h
@@ -119,8 +119,8 @@ void spinlock_init(spinlock_t* lock);
 void spin_lock(spinlock_t* lock);
 void spin_unlock(spinlock_t* lock);
 bool spin_trylock(spinlock_t* lock);
-void spin_lock_irqsave(spinlock_t* lock, uint32_t* flags);
-void spin_unlock_irqrestore(spinlock_t* lock, uint32_t flags);
+void spin_lock_irqsave(spinlock_t* lock, uint64_t* flags);
+void spin_unlock_irqrestore(spinlock_t* lock, uint64_t flags);
 
 // Mutex operations
 void mutex_init(mutex_t* mutex);
